Refactor: Remove Vuex, update wallet store and onboard
