+++ llm/README.md
@@ -208,8 +208,10 @@ yarn lint
    - ✅ Reverted to Bun 1.2.23 as intended (fixed CI lockfile issues)
    - ✅ Updated dependencies: vue-toastification, vue-ellipse-progress for Vue 3 compatibility
    - ✅ Added BlogPosting and FAQ schema markup to blog posts
-   - ✅ Removed unused Vue CLI configs and TypeScript utilities
-   - **Status**: 100% complete - Full Vue 3 migration with modern build tooling
+   - ✅ Removed unused Vue CLI configs, TypeScript utilities, and dead Vuex store code
+   - ✅ Fixed ESLint 9.x configuration with proper ES module support
+   - ✅ Updated PostCSS config for ES module compatibility
+   - **Status**: 100% complete - Full Vue 3 migration with modern build tooling and clean codebase
 2. **Security Vulnerabilities Fixed**: 
    - PostCSS vulnerability (moderate) - updated to >=8.4.31
    - ws vulnerability (high) - updated to >=8.18.0 via resolutions
